+++ mazikeen/Utils.py
@@ -13,7 +13,7 @@ def replaceVariables(line, dictReplVar, printer = Printer()):
     if line == None: return (True, line)
     searchStart = 0
     while (True):
-        m = re.search(r'(?<!\\)\${.*?}', line[searchStart:]) #SeNe optimize
+        m = re.search(r'(?<!\\)\${.*?}', line[searchStart:]) #ToDo: optimize
         if not m: 
             break
         foundVar = m.group()[2:-1]
@@ -61,24 +61,41 @@ def __getCompareLine(line, fh, compiledIgnoreLines):
     while(repeate):
         if not line: return line
         repeate = False
+        try:
+            strLine = line.decode('utf-8')
+        except:
+            return line
         for ignoreLine in compiledIgnoreLines:
-            if ignoreLine.match(line):
+            if ignoreLine.match(strLine):
                 line = fh.readline()
                 repeate = True
                 break
     return line
-    
+
+def normalizeEOL(line):
+    # Todo: can be optimized
+    if len(line) >=2 and line[-2:] == bytes('\r\n', 'utf-8'): 
+        line = line[:-2] + bytes('\n', 'utf-8')
+    elif len(line) >=1 and line[-1] == bytes('\r', 'utf-8'): 
+        line = line[:-1] + bytes('\n', 'utf-8')
+    return line
+
 def diffFiles(fileL, fileR, compiledIgnoreLines = [], binaryCompare = False):
-    fileFlags = "rb" if binaryCompare else "r"
-    with open(fileL, fileFlags) as fhL:
-        with open(fileR, fileFlags) as fhR:
+    with open(fileL, "rb") as fhL:
+        with open(fileR, "rb") as fhR:
             while True:
                 lineL = fhL.readline()
                 lineR = fhR.readline()
+                if (binaryCompare == False):
+                    lineL = normalizeEOL(lineL)
+                    lineR = normalizeEOL(lineR)
                 if (not lineL and not lineR): break
                 if lineL != lineR:
                     lineL = __getCompareLine(lineL, fhL, compiledIgnoreLines)
                     lineR = __getCompareLine(lineR, fhR, compiledIgnoreLines)
+                    if (binaryCompare == False):
+                        lineL = normalizeEOL(lineL)
+                        lineR = normalizeEOL(lineR)
                     if lineL != lineR:
                         return False
     return True